연습문제1

+++ 연습1/rectangle_calculator.py
@@ -1,4 +1,4 @@
-w = float(input("가로 길이를 입력하세요: "))
-h = float(input("세로 길이를 입력하세요: "))
+w = int(input("가로 길이를 입력하세요: "))
+h = int(input("세로 길이를 입력하세요: "))
 print("직사각형의 넓이:", w*h)
 print("직사각형의 둘레:", 2*(w+h))
